# internal/cw/types.go
// Package cw implements Morse sending: the queue, the pacing loops that keep a
// rig's CW buffer fed, and the local element generator used on rigs with no
// usable CAT CW buffer.
//
// # Canonical text format
//
// Text arriving at the API is in remoses canonical form, which is rig-neutral:
//
//   - Letters, digits and punctuation are sent as written. Case is not
//     significant.
//   - A prosign is written as '^' followed by the run of letters to be keyed
//     with no inter-character spacing: "^AR", "^BT", "^SK", "^KN".
//   - A '^' followed by anything that is not a letter is an error, so that a
//     stray caret is caught at the API boundary rather than mangled by the rig.
//
// Backends translate canonical form into their own encoding via
// backend.MorseSender.EncodeProsigns. The two target rigs differ completely:
// Icom uses '^' as a run-together marker natively, while Kenwood substitutes
// single ASCII symbols. Clients never see either dialect.
package cw

import (
	"errors"

	"github.com/hessu/remoses/internal/radio"
)

var (
	// ErrNotSupported is returned when a radio has no CW capability at all.
	ErrNotSupported = errors.New("cw: sending not supported by this radio")
	// ErrAborted is returned to callers waiting on a send that was cancelled.
	ErrAborted = errors.New("cw: aborted")
)

// CharError reports a character the target radio cannot key. The API turns it
// into a 422 naming the offending character and the accepted charset, because
// both target rigs silently mangle unsupported characters instead of
// complaining.
type CharError struct {
	Char    rune
	Offset  int
	Charset string
}

func (e *CharError) Error() string {
	return "cw: character " + string(e.Char) + " cannot be sent by this radio"
}

// Mode selects how Enqueue treats text already in the queue.
type Mode string

const (
	// Append adds to the end of the queue.
	Append Mode = "append"
	// Replace discards anything not yet handed to the radio, then appends.
	// It does not stop text already inside the rig's buffer; use Abort for that.
	Replace Mode = "replace"
)

// Sender is the uniform CW interface the API talks to, whether the Morse is
// generated by the rig from a CAT buffer or locally on a control line.
type Sender interface {
	// Enqueue validates text against Charset, translates prosigns, and adds it
	// to the queue. It returns the number of characters queued and does not
	// block waiting for them to be sent.
	Enqueue(text string, mode Mode) (queued int, err error)

	// Abort drops the local queue and stops the radio sending. Both halves
	// matter: up to a full buffer may already be inside the rig.
	Abort()

	// Status reports queue depth, busy state and speed.
	Status() radio.CWStatus

	// SetSpeed changes the keyer speed in words per minute.
	SetSpeed(wpm int) error

	// Charset is the set of characters this radio will key.
	Charset() string
}